Added quick test for stream to playground
