+++ Playgrounds/Example Languages.playground/timeline.xctimeline
@@ -3,17 +3,17 @@
    version = "3.0">
    <TimelineItems>
       <LoggerValueHistoryTimelineItem
-         documentLocation = "file:///Volumes/Personal/SPM/OysterKit/Playgrounds/Example%20Languages.playground#CharacterRangeLen=7&amp;CharacterRangeLoc=2773&amp;EndingColumnNumber=16&amp;EndingLineNumber=85&amp;StartingColumnNumber=9&amp;StartingLineNumber=85&amp;Timestamp=539029997.453246"
+         documentLocation = "file:///Volumes/Personal/SPM/OysterKit/Playgrounds/Example%20Languages.playground#CharacterRangeLen=7&amp;CharacterRangeLoc=1106&amp;EndingColumnNumber=16&amp;EndingLineNumber=43&amp;StartingColumnNumber=9&amp;StartingLineNumber=43&amp;Timestamp=539067696.83023"
          selectedRepresentationIndex = "0"
          shouldTrackSuperviewWidth = "NO">
       </LoggerValueHistoryTimelineItem>
       <LoggerValueHistoryTimelineItem
-         documentLocation = "file:///Volumes/Personal/SPM/OysterKit/Playgrounds/Example%20Languages.playground#CharacterRangeLen=10&amp;CharacterRangeLoc=1966&amp;EndingColumnNumber=23&amp;EndingLineNumber=66&amp;StartingColumnNumber=13&amp;StartingLineNumber=66&amp;Timestamp=539029997.453377"
+         documentLocation = "file:///Volumes/Personal/SPM/OysterKit/Playgrounds/Example%20Languages.playground#CharacterRangeLen=0&amp;CharacterRangeLoc=1090&amp;EndingColumnNumber=68&amp;EndingLineNumber=40&amp;StartingColumnNumber=68&amp;StartingLineNumber=40&amp;Timestamp=539067696.830422"
          selectedRepresentationIndex = "0"
          shouldTrackSuperviewWidth = "NO">
       </LoggerValueHistoryTimelineItem>
       <LoggerValueHistoryTimelineItem
-         documentLocation = "file:///Volumes/Personal/SPM/OysterKit/Playgrounds/Example%20Languages.playground#CharacterRangeLen=7&amp;CharacterRangeLoc=2304&amp;EndingColumnNumber=20&amp;EndingLineNumber=70&amp;StartingColumnNumber=13&amp;StartingLineNumber=70&amp;Timestamp=539029997.453458"
+         documentLocation = "file:///Volumes/Personal/SPM/OysterKit/Playgrounds/Example%20Languages.playground#CharacterRangeLen=0&amp;CharacterRangeLoc=1090&amp;EndingColumnNumber=68&amp;EndingLineNumber=40&amp;StartingColumnNumber=68&amp;StartingLineNumber=40&amp;Timestamp=539067696.830557"
          selectedRepresentationIndex = "0"
          shouldTrackSuperviewWidth = "NO">
       </LoggerValueHistoryTimelineItem>
